app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.1",
    "boto3>=1.36.26",
    "dynaconf>=3.2.10",
    "fastapi[standard]>=0.115.8",
    "llama-index>=0.12.19",
    "llama-index-embeddings-huggingface>=0.5.1",
    "llama-index-embeddings-instructor>=0.3.0",
    "llama-index-llms-ollama>=0.5.2",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.10",
    "python-json-logger>=3.2.1",
    "python-ulid>=3.0.0",
    "pyyaml>=6.0.2",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.38",
    "sqlalchemy-utils>=0.41.2",
    "uvicorn>=0.34.0",
]
